Use V4 binding for non-final properties where possible
When a property referenced in a binding is not marked as final, do not
automatically abort optimization.  Instead generate both V4 and V8
binidngs, and only fall back to the V8 binding if necessary at run time.

+++ src/qml/qml/v8/qv8bindings.cpp
@@ -296,7 +296,8 @@ QV8Bindings::configBinding(QObject *target, QObject *scope,
     rv->setNotifyOnValueChanged(true);
     rv->parent = this;
 
-    addref(); // This is decremented in Binding::destroy()
+    if (!i->isFallback)
+        addref(); // This is decremented in Binding::destroy()
 
     return rv;
 }
